requirements.txt
grpcio==1.7.0
pydub==0.20.0
setuptools==38.2.4
gapic-google-cloud-speech-v1
grpc-google-cloud-speech-v1
proto-google-cloud-speech-v1